chore: chaosctl physical-machine command
